Refactor helper functions to crate level
