+++ src/hinter/cwd_aware.rs
@@ -1,14 +1,10 @@
 use crate::{
+    hinter::get_first_token,
     history::SearchQuery,
     result::{ReedlineError, ReedlineErrorVariants::HistoryFeatureUnsupported},
     Hinter, History,
 };
 use nu_ansi_term::{Color, Style};
-use unicode_segmentation::UnicodeSegmentation;
-
-pub fn is_whitespace_str(s: &str) -> bool {
-    s.chars().all(char::is_whitespace)
-}
 
 /// A hinter that uses the completions or the history to show a hint to the user
 ///
@@ -68,22 +64,7 @@ impl Hinter for CwdAwareHinter {
     }
 
     fn next_hint_token(&self) -> String {
-        let mut reached_content = false;
-        let result: String = self
-            .current_hint
-            .split_word_bounds()
-            .take_while(|word| match (is_whitespace_str(word), reached_content) {
-                (_, true) => false,
-                (true, false) => true,
-                (false, false) => {
-                    reached_content = true;
-                    true
-                }
-            })
-            .collect::<Vec<&str>>()
-            .join("")
-            .to_string();
-        result
+        get_first_token(&self.current_hint)
     }
 }
